+++ NEWS.md
@@ -1,6 +1,7 @@
 # naomi 0.0.54
 
 * Add output stratification for age groups <1, 1-4.  This is simply disaggregated based on proportions from Spectrum with now uncertainty.
+* Add function `export_datapack()` to export an output package to datapack CSV format.
 
 # naomi 0.0.53